feat: Add tsqm-db binary

+++ examples/Helpers/DbHelper.php
@@ -2,8 +2,8 @@
 
 namespace Examples\Helpers;
 
+use Exception;
 use PDO;
-use Tsqm\TaskRepository;
 
 class DbHelper
 {
@@ -17,11 +17,17 @@ class DbHelper
     public function resetDb(string $table = "tsqm_tasks"): void
     {
         $driver = $this->pdo->getAttribute(PDO::ATTR_DRIVER_NAME);
-        $sql = TaskRepository::getCreateTableSql($table, $driver);
+        $code = 0;
+        $output = array();
+        exec(__DIR__."/../../bin/tsqm-db $driver $table", $output, $code);
+        $output = implode("\n", $output);
+        if ($code !== 0) {
+            throw new Exception("Failed to reset database: $output");
+        }
 
         $this->pdo->prepare("DROP TABLE IF EXISTS `$table`")->execute();
 
-        $queries = explode(";", $sql);
+        $queries = explode(";", $output);
         foreach ($queries as $query) {
             $query = trim($query);
             if ($query) {
